chore: improve tests
